+++ CHANGELOG.md
@@ -10,6 +10,17 @@
 ### Fixed
 -->
 
+## [0.2.1] — 2024-12-20
+
+### Security
+### Added
+Created a new `refine_parallel` method for the `SparseGrid` trait. The results in values being filled for the refined nodes in parallel.
+### Changed
+`RefinementFunctor` must now implement Send + Sync.
+### Deprecated
+### Removed
+Removed the useless sorting in the `refine` function. Presently we do not limit the number of nodes to refine in a given iteration.
+### Fixed
 
 ## [0.2.0] — 2024-12-16